pyproject.toml
[tool.poetry]
name = "rag-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "rag_test", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
haystack-ai = "2.2.4"
fastapi = "^0.111.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
